calculator.c file updated
input parameter given as float

+++ 3_Implementation/src/calculator_operations.c
@@ -1,21 +1,21 @@
 #include <calculator_operations.h>
 
-float add(int operand1, int operand2)
+float add(float operand1, float operand2)
 {
     return operand1 + operand2;
 }
 
-float subtract(int operand1, int operand2)
+float subtract(float operand1, float operand2)
 {
     return operand1 - operand2;
 }
 
-float multiply(int operand1, int operand2)
+float multiply(float operand1, float operand2)
 {
     return operand1 * operand2;
 }
 
-float divide(int operand1, int operand2)
+float divide(float operand1, float operand2)
 {
     if(0 == operand2)
         return 0;
